Move essay related components to a related package
