+++ src/app/app.module.ts
@@ -12,21 +12,25 @@ import { LoginComponent } from './login/login.component';
 import { UpdatePassComponent } from './update-pass/update-pass.component';
 import { DropdownDirective } from './shared/dropdown.directive';
 import { ProfileComponent } from './profile/profile.component';
-import { AuthGuardService } from './auth-guard.service';
 import { EditProfileComponent } from './edit-profile/edit-profile.component';
-import { AuthenticationService } from './services/authentication.service';
-import { UserService } from './services/user.service';
-
 
 import { HomeComponent } from './home/home.component';
 import { NavbarComponent } from './navbar/navbar.component';
 import { WorkstationComponent } from './workstation/workstation.component';
-import {CreateEssayComponent} from './create-essay/create-essay.component';
-import { EssayCardComponent } from './essay-card/essay-card.component';
+
+// Essay related components
+import {CreateEssayComponent} from './essay/create-essay/create-essay.component';
+import { ReviewEssay } from './essay/review-essay/review-essay.component';
+import { EssayCardComponent } from './essay/essay-card/essay-card.component';
+import { EditEssayComponent } from './essay/edit-essay/edit-essay.component';
+import { DeleteEssayComponent } from './essay/delete-essay/delete-essay.component';
+
+
+// Services
+import { AuthGuardService } from './auth-guard.service';
 import { EssayService } from './services/essay.service';
-import { EditEssayComponent } from './edit-essay/edit-essay.component';
-import { ReviewEssay } from './review-essay/review-essay.component';
-import { DeleteEssayComponent } from './delete-essay/delete-essay.component';
+import { AuthenticationService } from './services/authentication.service';
+import { UserService } from './services/user.service';
 
 // Loading environment
 import { environment } from '../environments/environment';
