+++ src/app/app.routing.module.ts
@@ -9,8 +9,8 @@ import { ProfileComponent } from './profile/profile.component';
 import { EditProfileComponent } from './edit-profile/edit-profile.component';
 import { HomeComponent } from './home/home.component';
 import { WorkstationComponent } from './workstation/workstation.component';
-import { EssayCardComponent } from './essay-card/essay-card.component';
-import { ReviewEssay } from './review-essay/review-essay.component';
+import { EssayCardComponent } from './essay/essay-card/essay-card.component';
+import { ReviewEssay } from './essay/review-essay/review-essay.component';
 import { EssayComponent } from './essay/essay.component';
 
 const routes: Routes = [
