+++ src/app/essay/create-essay/create-essay.component.scss
