+++ src/app/essay/create-essay/create-essay.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit, ChangeDetectorRef } from '@angular/core';
 import { FormGroup, FormBuilder, Validators } from '@angular/forms';
 
-import { EssayService } from '../services/essay.service';
-import { Essay } from '../../models/essay';
+import { EssayService } from '../../services/essay.service';
+import { Essay } from '../../../models/essay';
 
 @Component({
   selector: 'app-create-essay',
