+++ src/app/essay/delete-essay/delete-essay.component.html
