+++ src/app/essay/edit-essay/edit-essay.component.scss
