+++ src/app/essay/essay-card/essay-card.component.html
