+++ src/app/essay/essay-card/essay-card.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { EssayService } from '../services/essay.service';
-import { Essay } from '../../models/essay';
+import { EssayService } from '../../services/essay.service';
+import { Essay } from '../../../models/essay';
 
 @Component({
   selector: 'app-essay-card',
